Moved MarkdownToRtfConverter to right place. Minor changes after Copilot code review.
