+++ Rappen.AI.WinForm/ChatMessageLog.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Drawing;
 using System.Windows.Forms;
+using Rappen.XRM.Helpers.Extensions;
 
 namespace Rappen.AI.WinForm
 {
