+++ Rappen.AI.WinForm/Rappen.AI.WinForm.projitems
@@ -14,6 +14,5 @@
     <Compile Include="$(MSBuildThisFileDirectory)AIAppInsights.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ChatMessageHistory.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ChatMessageLog.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)MarkdownToRtfConverter.cs" />
   </ItemGroup>
 </Project>
